remove redundant module and files
