--- server/controller-sdk/src/main/java/ai/starwhale/mlops/domain/swds/index/SwdsDataLocation.java
@@ -1,54 +1,0 @@
-/*
- * Copyright 2022 Starwhale, Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package ai.starwhale.mlops.domain.swds.index;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * the location of the block data in one file
- */
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class SwdsDataLocation {
-
-    /**
-     * the offset to the file
-     */
-    int offset;
-
-    /**
-     * the byte size of this block
-     */
-    int size;
-
-    /**
-     * the file path where this block exists
-     */
-    String file;
-
-    static final String PATH_SPLITER = "/";
-
-    public void prependDsPath(String swdsPath) {
-        this.file = swdsPath + PATH_SPLITER + this.file;
-    }
-
-}
