+++ server/controller/pom.xml
@@ -14,10 +14,6 @@
   <dependencies>
     <dependency>
       <groupId>ai.starwhale</groupId>
-      <artifactId>starwhale-controller-sdk</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>ai.starwhale</groupId>
       <artifactId>storage-access-layer</artifactId>
     </dependency>
     <dependency>
