--- server/controller/src/main/java/ai/starwhale/mlops/domain/dataset/bo/DatasetIndex.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2022 Starwhale, Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package ai.starwhale.mlops.domain.dataset.bo;
-
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * index to make the access to SWDS Slices faster
- */
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class DatasetIndex {
-
-    /**
-     * storage file
-     */
-    String storagePath;
-
-    /**
-     * block items described by this index
-     */
-    List<DatasetBlock> datasetBlockList;
-
-    public DatasetIndex add(DatasetBlock datasetBlock) {
-        this.datasetBlockList.add(datasetBlock);
-        return this;
-    }
-
-
-}
