+++ server/controller/src/main/java/ai/starwhale/mlops/domain/system/agent/bo/Node.java
@@ -16,7 +16,6 @@
 
 package ai.starwhale.mlops.domain.system.agent.bo;
 
-import ai.starwhale.mlops.domain.node.Device;
 import ai.starwhale.mlops.domain.system.agent.AgentStatus;
 import java.util.List;
 import lombok.AllArgsConstructor;
