+++ server/controller/src/main/java/ai/starwhale/mlops/domain/task/bo/Task.java
@@ -16,7 +16,6 @@
 
 package ai.starwhale.mlops.domain.task.bo;
 
-import ai.starwhale.mlops.api.protocol.report.resp.ResultPath;
 import ai.starwhale.mlops.common.TimeConcern;
 import ai.starwhale.mlops.domain.job.step.bo.Step;
 import ai.starwhale.mlops.domain.task.status.TaskStatus;
