+++ server/controller/src/main/java/ai/starwhale/mlops/domain/task/converter/TaskBoConverter.java
@@ -16,9 +16,9 @@
 
 package ai.starwhale.mlops.domain.task.converter;
 
-import ai.starwhale.mlops.api.protocol.report.resp.ResultPath;
 import ai.starwhale.mlops.domain.job.step.bo.Step;
 import ai.starwhale.mlops.domain.system.agent.AgentConverter;
+import ai.starwhale.mlops.domain.task.bo.ResultPath;
 import ai.starwhale.mlops.domain.task.bo.Task;
 import ai.starwhale.mlops.domain.task.bo.TaskRequest;
 import ai.starwhale.mlops.domain.task.po.TaskEntity;
