+++ server/controller/src/main/java/ai/starwhale/mlops/domain/task/status/WatchableTask.java
@@ -16,9 +16,9 @@
 
 package ai.starwhale.mlops.domain.task.status;
 
-import ai.starwhale.mlops.api.protocol.report.resp.ResultPath;
 import ai.starwhale.mlops.domain.job.step.bo.Step;
 import ai.starwhale.mlops.domain.task.TaskWrapper;
+import ai.starwhale.mlops.domain.task.bo.ResultPath;
 import ai.starwhale.mlops.domain.task.bo.Task;
 import ai.starwhale.mlops.domain.task.bo.TaskRequest;
 import java.util.List;
